flask_app/requirements.txt
flask
uWSGI
pymongo
PyYAML